Cleanup the main folder
